Prepare the development environment
Installs gems for testing and development.

An anonymous volume is used to prevent the error 'A server is already running. Check /usr/src/app/tmp/pids/server.pid.' This error occurs when the container is not stopped gracefully. In development, the project folder is mounted into the container, so the PID is persisted. The anonymous volume will help to avoid this issue.

pry version is locked to prevent noisy warning messages: https://github.com/pry/pry/issues/2097.